backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
supabase==2.3.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
requests==2.31.0
aiofiles==23.2.1
boto3==1.28.85
botocore==1.31.85
joblib==1.3.2
watchtower==3.0.1